+++ src/Be.Vlaanderen.Basisregisters.GrAr.Contracts/BuildingRegistry/Legacy/BuildingWasMeasuredByGrb.cs
@@ -3,7 +3,7 @@ namespace Be.Vlaanderen.Basisregisters.GrAr.Contracts.BuildingRegistry
     using Contracts;
     using Common;
 
-    public class BuildingWasMeasuredByGrb : IQueueMessage
+    public sealed class BuildingWasMeasuredByGrb : IQueueMessage
     {
         public string BuildingId { get; }
         public string ExtendedWkbGeometry { get; }