+++ src/Be.Vlaanderen.Basisregisters.GrAr.Contracts/ParcelRegistry/Legacy/ParcelWasRegistered.cs
@@ -1,17 +1,16 @@
 namespace Be.Vlaanderen.Basisregisters.GrAr.Contracts.ParcelRegistry
 {
-    using System;
     using Contracts;
     using Common;
 
-    public class ParcelWasRegistered : IQueueMessage
+    public sealed class ParcelWasRegistered : IQueueMessage
     {
-        public Guid ParcelId { get; }
+        public string ParcelId { get; }
         public string VbrCaPaKey { get; }
         public Provenance Provenance { get; }
 
         public ParcelWasRegistered(
-            Guid parcelId,
+            string parcelId,
             string vbrCaPaKey,
             Provenance provenance)
         {