+++ src/Be.Vlaanderen.Basisregisters.GrAr.Contracts/ParcelRegistry/ParcelAddressWasDetachedV2.cs
@@ -1,26 +1,21 @@
 namespace Be.Vlaanderen.Basisregisters.GrAr.Contracts.ParcelRegistry
 {
-    using System;
     using Common;
 
     public sealed class ParcelAddressWasDetachedV2 : IQueueMessage
     {
-        public Guid ParcelId { get; }
-
-        public string CaPaKey { get; }
+        public string ParcelId { get; }
 
         public int AddressPersistentLocalId { get; }
 
         public Provenance Provenance { get; }
 
         public ParcelAddressWasDetachedV2(
-            Guid parcelId,
-            string caPaKey,
+            string parcelId,
             int addressPersistentLocalId,
             Provenance provenance)
         {
             ParcelId = parcelId;
-            CaPaKey = caPaKey;
             AddressPersistentLocalId = addressPersistentLocalId;
             Provenance = provenance;
         }